// src/types/users.ts
export interface StoreOwnerDataType {
  id: number;
  uid: string;
  name: string;
  email: string;
  email_verified_at: string;
  is_active: 0 | 1;
  created_at: string;
  updated_at: string;
}

export interface AdminsDataType {
  id: number;
  uid: string;
  name: string;
  email: string;
  email_verified_at: string;
  is_active: 0 | 1;
  created_at: string;
  updated_at: string;
}